moving some tests out
